Move to WalletWasabi.Fluent.MathNet namespace
